+++ src/crepe/Particle.h
@@ -24,13 +24,13 @@ public:
 	//! Accumulated force affecting the particle over time.
 	vec2 force_over_time;
 	//! Total lifespan of the particle in milliseconds.
-	uint32_t lifespan;
+	unsigned int lifespan;
 	//! Active state of the particle; true if it is in use, false otherwise.
 	bool active = false;
 	//! The time the particle has been alive, in milliseconds.
-	uint32_t time_in_life = 0;
+	unsigned int time_in_life = 0;
 	//! The angle at which the particle is oriented or moving.
-	double angle = 0;
+	float angle = 0;
 
 	/**
 	 * \brief Resets the particle with new properties.
@@ -43,7 +43,7 @@ public:
 	 * \param velocity  The initial velocity of the particle.
 	 * \param angle     The angle of the particle's trajectory or orientation.
 	 */
-	void reset(uint32_t lifespan, const vec2 & position, const vec2 & velocity, double angle);
+	void reset(unsigned int lifespan, const vec2 & position, const vec2 & velocity, float angle);
 	/**
 	 * \brief Updates the particle's state.
 	 *
